requirements.txt
Django==2.2.3
Pillow==5.2.0
psycopg2-binary==2.7.5
pymssql==2.1.4
PyPDF2==1.26.0
pytz==2018.5
beautifulsoup4==4.6.0
Jinja2==2.10
lxml
MarkupSafe==1.0
pkg-resources==0.0.0
